pyproject.toml
[project]
name = "sports-proj"
version = "0.1.0"
description = ""
authors = [
    {name = "Nestor Villa",email = "nestorvilla571@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]

[tool.poetry]
packages = [{include = "sports_proj", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
